+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gitaly changelog
 
+UNRELEASED
+
+- Vendor Gitlab::Git @ f9b946c1c9756533fd95c8735803d7b54d6dd204
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/563
+
 v0.74.0
 
 - Implement CreateRepositoryFromBundle RPC